--- lib/pages/WelcomePage.dart
@@ -1,64 +1,0 @@
-import 'package:flutter/material.dart';
-
-class WelcomePage extends StatefulWidget {
-  const WelcomePage({Key? key}) : super(key: key);
-
-  @override
-  State<WelcomePage> createState() => _WelcomePageState();
-}
-
-class _WelcomePageState extends State<WelcomePage> {
-  @override
-  Widget build(BuildContext context) {
-    return Center(
-      child: Container(
-        height: double.infinity,
-        width: double.infinity,
-        decoration: const BoxDecoration(
-          image: DecorationImage(
-              image: AssetImage('lib/assets/images/WelcomeBG.png'),
-              fit: BoxFit.cover),
-        ),
-        child: Padding(
-          padding: const EdgeInsets.only(top: 100),
-          child: Column(mainAxisAlignment: MainAxisAlignment.start, children: [
-            Image.asset('lib/assets/images/google_icon.png'),
-            Spacer(),
-            Padding(
-              padding: const EdgeInsets.only(bottom: 32),
-              child: Column(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: [
-                  Padding(
-                    padding: const EdgeInsets.all(12),
-                    child: SizedBox(
-                      height: 50,
-                      width: 300,
-                      child: ElevatedButton(
-                        onPressed: () {},
-                        child: const Text('GET STARTED'),
-                      ),
-                    ),
-                  ),
-                  SizedBox(
-                      height: 50,
-                      width: 300,
-                      child: ElevatedButton(
-                        style: ElevatedButton.styleFrom(
-                          primary: Colors.white,
-                        ),
-                        onPressed: () {},
-                        child: const Text('SIGN IN',
-                            style: TextStyle(
-                              color: Colors.black,
-                            )),
-                      ))
-                ],
-              ),
-            )
-          ]),
-        ),
-      ),
-    );
-  }
-}
